+++ StaticFields/headers/Control/ParticlePlacer.h
@@ -6,13 +6,13 @@
  * @param Particle* types: list of particles to be released,
  * @param int* numP: the number of particles for each type to be released,
  * @param int numTypes: the number of types of particles (length of types), 
- * @param long double distance: interparticle distance (for grid),
+ * @param double distance: interparticle distance (for grid),
  * @param Release type: type of particle release, 
  * @param Vector* position: For list release, gives position of each particle to be released, can be NULL for any non-list types,
  * @param Vector* velocities: similarly for list release, can be NULL even for list types, will default to 0 if null.
  * 
  * @return: Particle* array of particles released.
 */
-Particle* ParticleRelease(Particle* types, int* numP, int numTypes, long double distance, Release type, Vector* position, Vector* velocities);
+Particle* ParticleRelease(Particle* types, int* numP, int numTypes, double distance, Release type, Vector* position, Vector* velocities);
 
 #endif
