+++ StaticFields/headers/Model/Particle.h
@@ -6,11 +6,11 @@
 typedef struct Particle {
     char enabled; //Particles are disabled and not considered outside the relevant area, disabled Particles are cleared up every 100 steps.
     int charge; //in e
-    long double mass; //kg (sorry)
+    double mass; //kg (sorry)
     Vector position;
     Vector velocity;
     Vector acceleration; // required for the leapfrog algorithm. Starts at 0.
-    long double impactParameter; //for use in the air calculations
+    double impactParameter; //for use in the air calculations
 } Particle;
 #endif
